requirements.txt
crewai>=0.28.0
langchain>=0.1.0
langchain_openai>=0.0.5
langchain_community>=0.0.5
faiss-cpu>=1.7.4
openai>=1.12.0
fastapi>=0.109.0
uvicorn>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tiktoken>=0.5.0